Moved ClosedSplineCurve3 to legacy
